Title screen matches up with database.
When you load from title screen you will be able to select the specific roles. We now need to update the database to fit the characters in the game. All character currently load the Ranger model but this does not determine the stats. Spread out the layout of the title screen to prepare for a design overhaul.